Moved transaction tests to the transaction test suite.
