+++ modules/transactions/build.gradle
@@ -68,15 +68,16 @@ dependencies {
     integrationTestImplementation project(':ignite-metastorage-api')
     integrationTestImplementation project(':ignite-metastorage')
     integrationTestImplementation project(':ignite-network-api')
+    integrationTestImplementation project(':ignite-network')
     integrationTestImplementation project(':ignite-placement-driver-api')
     integrationTestImplementation project(':ignite-raft-api')
     integrationTestImplementation project(':ignite-raft')
     integrationTestImplementation project(':ignite-table')
     integrationTestImplementation project(':ignite-metrics')
-    integrationTestImplementation project(':ignite-low-watermark')
-    integrationTestImplementation project(':ignite-schema')
     integrationTestImplementation project(':ignite-configuration')
     integrationTestImplementation project(':ignite-configuration-root')
+    integrationTestImplementation project(':ignite-schema')
+    integrationTestImplementation project(':ignite-low-watermark')
     integrationTestImplementation libs.jetbrains.annotations
     integrationTestImplementation(testFixtures(project(':ignite-core')))
     integrationTestImplementation(testFixtures(project(':ignite-transactions')))
